Can add new tags when adding tags to claims, removed longclicklistener from manage tags. manage tags extends adding tags.
