+++ Team9Project/src/com/indragie/cmput301as1/ManageTagsActivity.java
@@ -16,17 +16,11 @@
  */
 package com.indragie.cmput301as1;
 
-import java.util.List;
-
-import android.app.ListActivity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.ActionMode;
 import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.AdapterView;
 
 
 /**
@@ -34,30 +28,15 @@ import android.widget.AdapterView;
  * Can direct user to activities for adding or editing tags. 
  * Allows user to remove existing tags. 
  */
-public class ManageTagsActivity extends ListActivity implements TypedObserver<List<Tag>>{
+public class ManageTagsActivity extends TagAddToClaimActivity{
 
 	//================================================================================
 	// Constants
 	//================================================================================
 	
-	private static final int ADD_TAG_REQUEST= 1;
 	private static final int EDIT_TAG_REQUEST= 2;
-	private static final String TAG_FILENAME = "tags";
 
 	public static final String EXTRA_TAG = "com.indragie.cmput301as1.EXTRA_TAG";
-	
-	//================================================================================
-	// Properties
-	//================================================================================
-	
-	/**
-	 * List model of tags.
-	 */
-	private ListModel<Tag> listModel;
-	/**
-	 * Index of a item that is long pressed.
-	 */
-	private int longPressedItemIndex;
 
 	//================================================================================
 	// Activity Callbacks
@@ -66,14 +45,10 @@ public class ManageTagsActivity extends ListActivity implements TypedObserver<Li
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		getActionBar().setDisplayHomeAsUpEnabled(true);
 		
-		listModel = new ListModel<Tag>(TAG_FILENAME, this);
-		listModel.addObserver(this);
-		setListAdapter(new TagArrayAdapter(this, listModel.getItems()));
+		setUpActionBarAndModel();
 		
-		
-		final ActionMode.Callback longClickCallback = new ActionMode.Callback() {
+		final ActionMode.Callback clickCallback = new ActionMode.Callback() {
 			@Override
 			public boolean onActionItemClicked(ActionMode mode, MenuItem item) {
 				switch (item.getItemId()) {
@@ -82,7 +57,7 @@ public class ManageTagsActivity extends ListActivity implements TypedObserver<Li
 						mode.finish();
 						return true;
 					case R.id.action_delete:
-						listModel.remove(longPressedItemIndex);
+						listModel.remove(pressedItemIndex);
 						mode.finish();
 						return true;
 					default:
@@ -106,44 +81,8 @@ public class ManageTagsActivity extends ListActivity implements TypedObserver<Li
 			}
 		};
 
-		getListView().setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
-			@Override
-			public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-				longPressedItemIndex = position;
-				startActionMode(longClickCallback);
-				return true;
-			}
-		});
-				
+		setUpItemClickListener(clickCallback);
 		
-	}
-	
-	@Override 
-	public boolean onCreateOptionsMenu(Menu menu){
-		super.onCreateOptionsMenu(menu);
-
-		MenuInflater inflater = getMenuInflater();
-		inflater.inflate(R.menu.contextual_add,menu);
-		return super.onCreateOptionsMenu(menu);
-	}
-	
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		switch (item.getItemId()) {
-		case R.id.action_add_tag:
-			startAddTagActivity();
-			return true;
-		case android.R.id.home:
-			finish();
-			return true;
-		default:
-			return super.onOptionsItemSelected(item);
-		}
-	}
-	
-	@Override
-	public void update(TypedObservable<List<Tag>> o, List<Tag> tags) {
-		setListAdapter(new TagArrayAdapter(this, tags));
 	}
 	
 	@Override
@@ -159,32 +98,15 @@ public class ManageTagsActivity extends ListActivity implements TypedObserver<Li
 	}
 	
 	//================================================================================
-	// Add/Edit a tag
+	// Edit a tag
 	//================================================================================
-	
-	/**
-	 * Starts the activity to add a tag.
-	 */
-	private void startAddTagActivity() {
-		Intent addTagIntent = new Intent(this, TagAddActivity.class);
-		startActivityForResult(addTagIntent, ADD_TAG_REQUEST);
-	}
-	
-	/**
-	 * Adds a tag to the list model from a resulting activity.
-	 * @param data The intent from resulting activity.
-	 */
-	private void onAddTag(Intent data) {
-		Tag tag = (Tag)data.getSerializableExtra(TagAddActivity.ADDED_TAG);
-		listModel.add(tag);
-	}
 	
 	/**
 	 * Starts the activity to edit a tag.
 	 */
 	private void startEditTagActivity() {
 		Intent editTagIntent = new Intent(this, TagEditActivity.class);
-		editTagIntent.putExtra(EXTRA_TAG, listModel.getItems().get(longPressedItemIndex));
+		editTagIntent.putExtra(EXTRA_TAG, listModel.getItems().get(pressedItemIndex));
 		startActivityForResult(editTagIntent, EDIT_TAG_REQUEST);
 	}
 	
@@ -194,7 +116,7 @@ public class ManageTagsActivity extends ListActivity implements TypedObserver<Li
 	 */
 	private void onEditTag(Intent data) {
 		Tag tag = (Tag)data.getSerializableExtra(TagAddActivity.ADDED_TAG);
-		listModel.set(longPressedItemIndex, tag);
+		listModel.set(pressedItemIndex, tag);
 	}
 
 }
